requirements.txt
Flask~=2.0.2
Pillow~=8.4.0